# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'Eden Allergies API Documentation'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
paths:
...
  /api/allergies:
    get:
      summary: 'GET api/allergies'
      description: 'Display a listing of allergies'
      parameters:
        -
          in: query
          name: page
          description: 'Page number to show.'
          example: 18
          required: false
          schema:
            type: integer
            description: 'Page number to show.'
            example: 18
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: null
                      title: null
                      description: null
                      created_at: null
                      updated_at: null
                      deleted_at: null
                      links:
                        meals: 'http://localhost:8000/api/allergies//meals/'
                    -
                      id: null
                      title: null
                      description: null
                      created_at: null
                      updated_at: null
                      deleted_at: null
                      links:
                        meals: 'http://localhost:8000/api/allergies//meals/'
                  links:
                    first: '/?page=1'
                    last: '/?page=1'
                    prev: null
                    next: null
                  meta:
                    current_page: 1
                    from: 1
                    last_page: 1
                    links:
                      -
                        url: null
                        label: '&laquo; Previous'
                        active: false
                      -
                        url: '/?page=1'
                        label: '1'
                        active: true
                      -
                        url: null
                        label: 'Next &raquo;'
                        active: false
                    path: /
                    per_page: '10'
                    to: 2
                    total: 2
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: null
                        title: null
                        description: null
                        created_at: null
                        updated_at: null
                        deleted_at: null
                        links:
                          meals: 'http://localhost:8000/api/allergies//meals/'
                      -
                        id: null
                        title: null
                        description: null
                        created_at: null
                        updated_at: null
                        deleted_at: null
                        links:
                          meals: 'http://localhost:8000/api/allergies//meals/'
                    items:
                      type: object
                  links:
                    type: object
                    example:
                      first: '/?page=1'
                      last: '/?page=1'
                      prev: null
                      next: null
                  meta:
                    type: object
                    example:
                      current_page: 1
                      from: 1
                      last_page: 1
                      links:
                        -
                          url: null
                          label: '&laquo; Previous'
                          active: false
                        -
                          url: '/?page=1'
                          label: '1'
                          active: true
                        -
                          url: null
                          label: 'Next &raquo;'
                          active: false
                      path: /
                      per_page: '10'
                      to: 2
                      total: 2
      tags:
        - 'Allergy Endpoints'
      security: []
    post:
      summary: 'POST api/allergies'
      description: 'Store a newly created allergy in storage.'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: null
                    title: null
                    description: null
                    created_at: null
                    updated_at: null
                    deleted_at: null
                    links:
                      meals: 'http://localhost:8000/api/allergies//meals/'
                properties:
                  data:
                    type: object
                    example:
                      id: null
                      title: null
                      description: null
                      created_at: null
                      updated_at: null
                      deleted_at: null
                      links:
                        meals: 'http://localhost:8000/api/allergies//meals/'
      tags:
        - 'Allergy Endpoints'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 'The title of the allergy'
                  example: debitis
                description:
                  type: string
                  description: 'The description of the allergy'
                  example: deserunt
              required:
                - title
...
  /api/users/allergies:
    get:
      summary: 'GET  api/users/allergies'
      description: 'Get allergies for authenticated user'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    -
                      id: null
                      title: null
                      description: null
                      created_at: null
                      updated_at: null
                      deleted_at: null
                      links:
                        meals: 'http://localhost:8000/api/allergies//meals/'
                    -
                      id: null
                      title: null
                      description: null
                      created_at: null
                      updated_at: null
                      deleted_at: null
                      links:
                        meals: 'http://localhost:8000/api/allergies//meals/'
                properties:
                  data:
                    type: array
                    example:
                      -
                        id: null
                        title: null
                        description: null
                        created_at: null
                        updated_at: null
                        deleted_at: null
                        links:
                          meals: 'http://localhost:8000/api/allergies//meals/'
                      -
                        id: null
                        title: null
                        description: null
                        created_at: null
                        updated_at: null
                        deleted_at: null
                        links:
                          meals: 'http://localhost:8000/api/allergies//meals/'
                    items:
                      type: object
      tags:
        - 'User Endpoints'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []